# api/openapi.yaml
openapi: 3.1.0
info:
  title: SYNC Server API
  version: 0.1.0
  description: |
    Federated, privacy-focused 1:1 chat server.
    Access tokens are short-lived HS256 JWTs (1 h TTL).
    Refresh tokens are 30-day opaque tokens subject to rotation and family-replay detection.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      required: [id, sender_id, recipient_id, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        recipient_id:
          type: string
          format: uuid
        content:
          type: string
        delivered_at:
          type: string
          format: date-time
          nullable: true
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    SendMessageRequest:
      type: object
      required: [recipient_id, content]
      properties:
        recipient_id:
          type: string
          description: |
            Local conversation partner UUID, or remote partner user ID when
            `recipient_server_url` is provided.
        recipient_server_url:
          type: string
          format: uri
          nullable: true
          description: |
            Required for cross-server send to a user on another planet.
            Message path: device A -> server A -> target server -> target device.
        content:
          type: string
          minLength: 1
    ResolveContactRequest:
      type: object
      required: [recipient_id, recipient_server_url]
      properties:
        recipient_id:
          type: string
          description: Target user ID on the destination server.
        recipient_server_url:
          type: string
          format: uri
    ResolveContactResponse:
      type: object
      required: [partner_id, recipient_id, recipient_server_url, display_handle]
      properties:
        partner_id:
          type: string
          format: uuid
          description: Local partner UUID to use for `/api/messages/{partner_id}` conversation APIs.
        recipient_id:
          type: string
        recipient_server_url:
          type: string
          format: uri
        display_handle:
          type: string
    GuildSnapshot:
      type: object
      required:
        - active_days
        - level
        - contribution_score
        - rank
        - level_progress_percent
        - daily_outbound_messages_enforced
        - daily_outbound_messages_sent
        - daily_attachment_sends_enforced
        - daily_attachment_sends_sent
        - allowed_attachment_types
        - daily_friend_adds_enforced
        - daily_friend_adds_sent
        - challenge_state
      properties:
        active_days:
          type: integer
        level:
          type: integer
          minimum: 1
          maximum: 10
        contribution_score:
          type: integer
        rank:
          type: string
          enum: [F, E, D, C, B, A, S]
        next_level_active_days:
          type: integer
          nullable: true
        level_progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        daily_outbound_messages_enforced:
          type: boolean
        daily_outbound_messages_limit:
          type: integer
          nullable: true
        daily_outbound_messages_sent:
          type: integer
        daily_outbound_messages_remaining:
          type: integer
          nullable: true
        daily_attachment_sends_enforced:
          type: boolean
        daily_attachment_send_limit:
          type: integer
          nullable: true
        daily_attachment_sends_sent:
          type: integer
        daily_attachment_sends_remaining:
          type: integer
          nullable: true
        allowed_attachment_types:
          type: array
          items:
            type: string
        daily_friend_adds_enforced:
          type: boolean
        daily_friend_add_limit:
          type: integer
          nullable: true
        daily_friend_adds_sent:
          type: integer
        daily_friend_adds_remaining:
          type: integer
          nullable: true
        challenge_state:
          type: string
          enum: [none, challenged, frozen]
          description: >
            Client-visible progression challenge state.
            "none" – no challenge active.
            "challenged" – suspicious activity detected; progression paused pending clean activity.
            "frozen" – repeated suspicious attempts; progression frozen until review or quiet-period clears.
        pending_milestone_notification:
          allOf:
            - $ref: '#/components/schemas/GuildMilestoneNotification'
          nullable: true
          description: >
            Non-null when a level-up, rank-up, or attachment-type unlock was detected on this
            request.  The client is responsible for dismissing the notification locally; the server
            does not persist per-user notification state.  Safe to ignore.
    GuildMilestoneNotification:
      type: object
      description: >
        Payload the client uses to render a milestone toast, badge, or unlock animation.
        Returned as `pending_milestone_notification` in `GuildSnapshot` when a server-side
        progression event was detected on this request.  Null when no milestone is pending.
        The client is responsible for dismissing / persisting the notification locally;
        the server does not maintain per-user notification state.
      required:
        - kind
        - badge_label
        - headline_key
        - detail_key
        - new_value
      properties:
        kind:
          type: string
          enum: [level_up, rank_up, unlock_attachment_type]
          description: >
            Kind of progression milestone that triggered this notification.
            "level_up" – the user reached a new guild level on this request.
            "rank_up" – the user's contribution score crossed a rank threshold.
            "unlock_attachment_type" – a new attachment MIME type became available.
        badge_label:
          type: string
          description: Human-readable badge label, e.g. "Level 5" or "Rank D".
        headline_key:
          type: string
          description: Localisation key for the unlock headline, e.g. "guild.milestone.level_up".
        detail_key:
          type: string
          description: Localisation key for the unlock sub-copy, e.g. "guild.milestone.level_5_detail".
        unlocked_value:
          type: string
          nullable: true
          description: >
            Optional specific thing unlocked, e.g. "application/pdf" for an attachment-type unlock.
            Null for level-up and rank-up milestones.
        new_value:
          type: string
          description: >
            New level or rank value after the milestone, e.g. "5" for a level-up or "D" for a rank-up.
...
    UnreadCounts:
      type: object
      description: Map of partner_user_id (UUID string) → unread message count
      additionalProperties:
        type: integer
        format: int64
      example:
        "550e8400-e29b-41d4-a716-446655440000": 3
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
    TrustLimitExceededResponse:
      type: object
      description: >
        Returned with HTTP 429 when a guild-level action cap is reached.
        The `guild` snapshot lets clients explain the limit and when it resets.
      required: [error, code, retry_after_seconds, guild]
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          enum:
            - daily_message_limit_reached
            - friend_add_limit_exceeded
          description: Machine-readable error code.
        retry_after_seconds:
          type: integer
          description: Seconds until the daily counter resets (start of next UTC day).
        guild:
          $ref: '#/components/schemas/GuildSnapshot'
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
paths:
...
  /api/messages:
    post:
      operationId: sendMessage
      summary: Send a direct message
      tags: [messages]
      description: |
        Persists the message to the sender's local server first.

        - Local target: server delivers directly to recipient device.
        - Cross-server target (when `recipient_server_url` is provided):
          server A relays to server B, and server B delivers to the recipient device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Recipient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          description: Daily outbound message limit reached for the sender's current guild level.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustLimitExceededResponse'
  /api/messages/resolve-contact:
    post:
      operationId: resolveContact
      summary: Resolve friend target into a local partner conversation ID
      tags: [messages]
      description: |
        Use this before starting a cross-server chat. Client provides
        `recipient_id` and `recipient_server_url`; server responds with a local
        `partner_id` UUID usable with conversation endpoints.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveContactRequest'
      responses:
        '200':
          description: Contact resolved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolveContactResponse'
        '400':
          description: Invalid target ID or server URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Daily friend-add limit reached for the sender's current guild level.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustLimitExceededResponse'
  /api/messages/unread-counts:
    get:
      operationId: getUnreadCounts
      summary: Get unread message counts per conversation partner
      tags: [messages]
      responses:
        '200':
          description: Map of partner UUID → unread count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCounts'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/messages/{partner_id}:
    get:
      operationId: getConversation
      summary: List messages in a conversation (keyset-paginated)
      tags: [messages]
      description: |
        Returns up to `limit` messages (max 100) ordered newest-first.
        Pass `before` (a message UUID) to fetch the page preceding that message.
      parameters:
        - name: partner_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: before
          in: query
          required: false
          description: Message UUID cursor — return messages older than this one
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Page of messages (newest-first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/messages/{partner_id}/read:
    post:
      operationId: markMessagesRead
      summary: Mark all unread messages from a partner as delivered and read
      tags: [messages]
      parameters:
        - name: partner_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Number of messages updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...